Move Plots.jl recipes into a package extension
This reduces the load time on Julia 1.12 from ~0.2s to ~0.17s.
